pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tablemaster"
version = "2.1.0"
description = "tablemaster is a Python toolkit for moving and managing tabular data across databases, Feishu/Lark, Google Sheets, and local files with one consistent API."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Livid", email = "livid.su@gmail.com" }
]
dependencies = [
    "pandas>=1.5,<3",
    "pyyaml>=6",
    "python-dateutil>=2.8",
    "tqdm>=4.60",
    "typer>=0.12,<1",
]

[project.optional-dependencies]
database = ["PyMySQL>=1.1", "SQLAlchemy>=2.0"]
mysql = ["PyMySQL>=1.1", "SQLAlchemy>=2.0"]
feishu = ["requests>=2.28"]
gspread = ["gspread>=6"]
local = ["openpyxl>=3.1", "pyarrow>=14"]
schema = ["SQLAlchemy>=2.0", "PyMySQL>=1.1", "psycopg2-binary>=2.9"]
all = [
    "tablemaster[database,feishu,gspread,local,schema]",
]

[project.urls]
Homepage = "https://github.com/ilivid/tablemaster"

[project.scripts]
tablemaster = "tablemaster.cli:app"

[tool.setuptools.packages.find]
include = ["tablemaster*"]
